+++ lib/src/main/java/org/asamk/signal/manager/helper/IncomingMessageHandler.java
@@ -257,7 +257,7 @@ public final class IncomingMessageHandler {
         var notAllowedToSendToGroup = isNotAllowedToSendToGroup(envelope, content);
         final var groupContext = getGroupContext(content);
         if (groupContext != null && groupContext.getGroupV2().isPresent()) {
-            handleGroupV2Context(groupContext.getGroupV2().get());
+            handleGroupV2Context(groupContext.getGroupV2().get(), receiveConfig.ignoreAvatars());
         }
         // Check again in case the user just joined the group
         notAllowedToSendToGroup = notAllowedToSendToGroup && isNotAllowedToSendToGroup(envelope, content);
@@ -377,15 +377,12 @@ public final class IncomingMessageHandler {
 
         if (content.getStoryMessage().isPresent()) {
             final var message = content.getStoryMessage().get();
-            actions.addAll(handleSignalServiceStoryMessage(message, sender, receiveConfig.ignoreAttachments()));
+            actions.addAll(handleSignalServiceStoryMessage(message, sender, receiveConfig));
         }
 
         if (content.getSyncMessage().isPresent()) {
             var syncMessage = content.getSyncMessage().get();
-            actions.addAll(handleSyncMessage(envelope,
-                    syncMessage,
-                    senderDeviceAddress,
-                    receiveConfig));
+            actions.addAll(handleSyncMessage(envelope, syncMessage, senderDeviceAddress, receiveConfig));
         }
 
         return actions;
@@ -499,7 +496,7 @@ public final class IncomingMessageHandler {
             if (message.getStoryMessage().isPresent()) {
                 actions.addAll(handleSignalServiceStoryMessage(message.getStoryMessage().get(),
                         sender.recipientId(),
-                        receiveConfig.ignoreAttachments()));
+                        receiveConfig));
             }
         }
         if (syncMessage.getRequest().isPresent() && account.isPrimaryDevice()) {
@@ -525,7 +522,9 @@ public final class IncomingMessageHandler {
             try {
                 final var groupsMessage = syncMessage.getGroups().get();
                 context.getAttachmentHelper()
-                        .retrieveAttachment(groupsMessage, input -> context.getSyncHelper().handleSyncDeviceGroups(input, receiveConfig.ignoreAvatars()));
+                        .retrieveAttachment(groupsMessage,
+                                input -> context.getSyncHelper()
+                                        .handleSyncDeviceGroups(input, receiveConfig.ignoreAvatars()));
             } catch (Exception e) {
                 logger.warn("Failed to handle received sync groups, ignoring: {}", e.getMessage());
             }
@@ -553,7 +552,8 @@ public final class IncomingMessageHandler {
                 final var contactsMessage = syncMessage.getContacts().get();
                 context.getAttachmentHelper()
                         .retrieveAttachment(contactsMessage.getContactsStream(),
-                                input -> context.getSyncHelper().handleSyncDeviceContacts(input, receiveConfig.ignoreAvatars()));
+                                input -> context.getSyncHelper()
+                                        .handleSyncDeviceContacts(input, receiveConfig.ignoreAvatars()));
             } catch (Exception e) {
                 logger.warn("Failed to handle received sync contacts, ignoring: {}", e.getMessage());
             }
@@ -758,9 +758,9 @@ public final class IncomingMessageHandler {
                                 groupV1 = new GroupInfoV1(groupId);
                             }
 
-                            if (groupInfo.getAvatar().isPresent()) {
+                            if (groupInfo.getAvatar().isPresent() && !receiveConfig.ignoreAvatars()) {
                                 var avatar = groupInfo.getAvatar().get();
-                                context.getGroupHelper().downloadGroupAvatar(groupV1.getGroupId(), avatar, receiveConfig.ignoreAvatars());
+                                context.getGroupHelper().downloadGroupAvatar(groupV1.getGroupId(), avatar);
                             }
 
                             if (groupInfo.getName().isPresent()) {
@@ -800,7 +800,7 @@ public final class IncomingMessageHandler {
                 }
             }
             if (groupContext.getGroupV2().isPresent()) {
-                handleGroupV2Context(groupContext.getGroupV2().get());
+                handleGroupV2Context(groupContext.getGroupV2().get(), receiveConfig.ignoreAvatars());
             }
         }
 
@@ -882,7 +882,8 @@ public final class IncomingMessageHandler {
                 account.getStickerStore().addStickerPack(sticker);
             }
             if (!receiveConfig.ignoreStickers()) {
-                context.getJobExecutor().enqueueJob(new RetrieveStickerPackJob(stickerPackId, messageSticker.getPackKey()));
+                context.getJobExecutor()
+                        .enqueueJob(new RetrieveStickerPackJob(stickerPackId, messageSticker.getPackKey()));
             }
         }
         return actions;
@@ -895,14 +896,14 @@ public final class IncomingMessageHandler {
     private List<HandleAction> handleSignalServiceStoryMessage(
             SignalServiceStoryMessage message,
             RecipientId source,
-            boolean ignoreAttachments
+            ReceiveConfig receiveConfig
     ) {
         var actions = new ArrayList<HandleAction>();
         if (message.getGroupContext().isPresent()) {
-            handleGroupV2Context(message.getGroupContext().get());
+            handleGroupV2Context(message.getGroupContext().get(), receiveConfig.ignoreAvatars());
         }
 
-        if (!ignoreAttachments) {
+        if (!receiveConfig.ignoreAttachments()) {
             if (message.getFileAttachment().isPresent()) {
                 context.getAttachmentHelper().downloadAttachment(message.getFileAttachment().get());
             }
@@ -924,13 +925,14 @@ public final class IncomingMessageHandler {
         return actions;
     }
 
-    private void handleGroupV2Context(final SignalServiceGroupV2 groupContext) {
+    private void handleGroupV2Context(final SignalServiceGroupV2 groupContext, final boolean ignoreAvatars) {
         final var groupMasterKey = groupContext.getMasterKey();
 
         context.getGroupHelper()
                 .getOrMigrateGroup(groupMasterKey,
                         groupContext.getRevision(),
-                        groupContext.hasSignedGroupChange() ? groupContext.getSignedGroupChange() : null);
+                        groupContext.hasSignedGroupChange() ? groupContext.getSignedGroupChange() : null,
+                        ignoreAvatars);
     }
 
     private void handleIncomingProfileKey(final byte[] profileKeyBytes, final RecipientId source) {