+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## [Unreleased]
+
 ## [0.14.0] - 2026-03-01
 
 **Attention**: Now requires Java 25